Add a unified comment header to each relevant source file
In the beginning of each source file there should be a unified header with
some basic licensing information as well as the authorship of that specific
file. The authorship is adopted from the Doxygen comment, which was already
present. This is one step to get the project completely compliant with the
GPL instructions given at [1].

[1] https://www.gnu.org/licenses/gpl-howto.html