Remove unused function "dayofweek"
The function "dayofweek" is not compiled due to some conditional statement,
which always returns false. It was planned for later use, however it actually
isn't needed (classic example for YAGNI).

This function is currently not used and probably never will be. In order to
keep things clean and simple remove it for now.