Remove support for the TIX display
It probably was developed just for the fun of it and never really
tested, because it was basically never built by anyone. However there
are some problems reported with it, see [1].

There are other issues with the TIX display in its current form anyway:

    - No support for RGB LEDs
    - No frontpanels are available
    - No schematics are available
    - Only very few things in common with the Wordclock display itself,
      as the LEDs in the Wordclock are controlled word wise, whereas the
      LEDs in the TIX display are controlled individually.

Remove the support for it completely will make the maintenance of the
source code easier in the future.

[1] https://www.mikrocontroller.net/topic/156661#2420410