Add macros for easier handling of port definitions
Right now ports are defined with up to four macros, e.g.:

    #define WHATEVER_PORT	PORTD
    #define WHATEVER_DDR	DDRD
    #define WHATEVER_PIN	PIND
    #define WHATEVER_BIT	4

This is quite redundant and error prone. With these macros a single
define like the following is sufficient:

    #define WHATEVER PORTD, 4

With the macros provided here, namely PORT(), DDR(), PIN() and BIT()
access to all registers is possible, just like in the case above.