dcf77.c: Remove _FLAG_COUNT enumerator
This was meant to be used to output an error when there are too many
enumerators within the enumeration. However this can't work like this,
as the preprocessor actually doesn't know the value for _FLAG_COUNT.

Therefore the flag including the "error logic" gets removed. However a
warning is added to the Doxygen comment block describing the problem
with more than eight enumerations.