dcf77.c: Replace various occurrences of module by receiver
To make things clearer whenever referring to the module demodulating the
DCF77 signal it is now called "receiver", whereas when talking about the
unit of compilation we now use "module" consequently. On top of that
replace various occurrences of "high active" to "active high" as well as
"low active" to "low active".