wceeprom: Move wcEeprom_getData() from wceeprom.h to wceeprom.c
The latter file contains the actual implementation for the whole module,
so it makes more sense to put this function there, too. This saves fifty
bytes, which is another reason for the move.

On top of that it makes no sense to declare this function as inline, as
this can be automatically achieved by various optimizations the compiler
is doing.