display_wc_ger.c: Put definition of langMode at top
This puts the definition of the variable more at the top. It is used in
both cases (DISPLAY_DEACTIVATABLE_ITIS enabled and/or disabled), so it
makes more sense to define it above the if paragraph. However in case of
DISPLAY_DEACTIVATABLE_ITIS == 1, the value must be divided by two.