display_wc_ger3.c: Rename DISP_SETBIT() to DISP_SETBIT_MIN()
As there is another macro similar to this one dealing with hours, this
one should be renamed, so both of them are unique. This macro can only
deal with the minute part of a display state anyway as it subtracts
DWP_MIN_FIRST from the argument given, before doing the bit shift.