uart.c: Fix an issue with wrong values being assigned to the UBRR0 register
f992e20956d71e69015f5e26ee00eba2b461de0b introduced a bug, so that the wrong
values were assigned to the UBRR registers, which rendered the UART useless.

The error was that the calculated value for the UBRR (16-bit) was assigned to
both the upper (UBRR0H) as well as the lower (UBRR0L) bytes. This fixes it by
simply assigning the calculated value (UBRR_VALUE) directly to the UBRR0
register.