usermodes.c: Enter "hue fading" mode AutoHueState_enter() correctly
This makes sure that the "hue fading" mode is entered correctly by removing
the call of the "ISR", which will be executed by the appropriate dispatcher
anyway. It sets up the counter with the proper value instead. Furthermore
the logic of the counter is inverted, so it is actually counting upwards,
which is more intuitive to understand.