user: Rename e_userCommands to user_command_t and switch prefix
This renames the enumeration representing all of the  possible commands from
the user from e_userCommands to user_command_t, which makes it more clear
that it is an actual type.

Besides just renaming the enumeration, also the prefix for all of the commands
is now "UC_" instead of "UI_". The appropriate Doxygen comments are updated,
too.