user: Rename various functions from *_HandleIr() to *_HandleUserCommand()
As this functions support user commands in general and are not specific to IR
commands anymore, it makes sense to rename them appropriately. Furthermore the
type of the parameter for all of these functions is changed to "user_command_t"
and the name of the parameter is changed to "command". The Doxygen comments are
updated, too.