uart: Rework module to work asynchronously and add receiving functionality
This reworks the UART module in a way that it works completely asynchronously
using interrupts. Some means to synchronize are provided, too. Furthermore it
also adds receiving functionality, which enables to get data received by the
UART hardware.

The reset functionality along with the macros BOOTLOADER_RESET_UART and
BOOTLOADER_RESET_WDT is removed by this, as it needs to be reimplemented
in the proper place later on.