ldr: ldr_init(): Enable ADC interrupt at the very end
The first measurement is taken directly, without the ISR being involved.
Therefore the ADC interrupt can only be enabled when the module was
initialized successfully.

It worked fine with the current setup beforehand, only because the
global interrupt flag was disabled when ldr_init() was called. This
makes sure it will also work when this is not the case.