datetime.h: Change definition of datetime_t::WD
This changes the definition of datetime_t so that it actually ranges
from 1 to 7, where 1 represents Monday. This definition makes sense as
it is more consistent with the DCF77 signal as well as the other fields.

This also fixes an issue where the conversion was actually messing this
field up, so that the value written to the RTC was invalid. This was
reported at [1].

[1]: https://www.mikrocontroller.net/topic/156661?goto=3489339#3489178