user: Change UserEepromParams::colorPresets to expect values up to 255
UserEepromParams::colorPresets expected an index to the actual step
value. As the steps itself are a detail of the PWM module itself,
it makes more sense to define these properties in a way that it
contains RGB colors itself, so all values from 0 up to 255 are valid.
All functions handling with these properties are changed, so they
are consistent with the new definition.