bootloader: wordboot: wordboot.c: Define buff as real variable
As the .bss section is used by mcusr anyway, it is more intuitive to define
buff as real variable also, rather than some tricky define. While this
increases the generated code by a couple of bytes, it increases readibility
and is less dependent on the compiler placing things at the right location.