src: log.c: Put string arrays into program space
This puts the string arrays into program space, which saves RAM space. Due to
limitations of the compiler this needs to be done in two stages, i.e. first
the string itself is defined and put in program space. The arrays themself only
contain pointers to these strings (type PGM_P).