src: preferences: Change signature of preferences_save()
This changes the signature of preferences_save() and makes it easier to use as
the start and length parameter are no longer required. As reading values from
the EEPROM and comparing them to the copy hold in SRAM is a cheap operation,
this is not too big of a performance hit. Furthermore the function is called
only seldom and it makes it easier for the callee if he does not need to keep
track of what was actually changed.