src: Implement module specifying format strings
This introduces a new module, which specifies various format strings and makes
them accessible througout the whole project. These strings are put into program
space once and can be re-used across different other modules.

On top of that this commit makes also sure that all of the existing printf()
invocations make use of this new facility.