src: Get rid of dedicated format module
This gets rid of the module dedicated for format strings, as it is not flexible
enough. In case these format specifiers are needed within a string, the string
would have to be built up dynamically during runtime using strcat(), which does
not make too much sense. Mixing these variables with hardcoded strings is not
a good either, so we have to live with the overhead introduced by specifying
strings over and over again.